backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Django backend with JWT auth and MCP tools management"
# readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0.0",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "psycopg2-binary>=2.9.9",
    "python-decouple>=3.8",
    "django-cors-headers>=4.3.0",
    "pillow>=10.0.0",
    "drf-spectacular>=0.28.0",
    "cryptography>=45.0.6",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.4.0",
    "pytest-django>=4.7.0",
    "factory-boy>=3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"*/migrations/*" = ["E501"]

[tool.ruff.isort]
known-first-party = ["backend"]
